Fix WSGIResponse on Python 3
HTTP body must be bytes. Don't check if __iter__() method to check if content
is a list or tuple, because bytes and str now have this method on Python 3.
Instead, check explicitly for bytes and str types using the six module.

Port wsgiwrappers test to Python 3, fix bytes/unicode issues:

* HTTP body must be bytes
* Filenames are native strings, request ensures that filenames are unicode