Fix proxy test
The proxy test relied on pythonpaste.org being up and existing.
It is neither. So instead we use httpbin.org which is often used
for this kind of thing. Unfortunately httpbin is now a react app
which means a lot of the HTML is generated client-side, which
means we need to choose wisely.

As the original comments indicate, the test is not particularly
robust and remains so.