Remove use of _lock around get and set timeout of SSL connection
The maintainers of pyOpenSSL have deprecated the tsafe module
because it is neither used nor complete. It was removed from
in paste in 78dd2ec01384 but because of a lack of test coverage
it caused a regression as described in #19.

Reviewing the code and pyOpenSSL, it appears that the right
fix is to simply not subclass the SSL.Connection, and let
get and set timeout be called on the socket object that
Connection keeps, without a lock.

A concern here is that if someone is doing something multi-
threaded with paste/httpserver the removal of the use of tsafe
may present some issues. On the other hand, if someone is
doing something multi-threaded, one hopes they are using a
more robust web and modern web server.

Fixes #19