Revert "Remove use of OpenSSL.tsafe, which links to OpenSSL.SSL anyways. (#16)"
This reverts commit 78dd2ec0138467305f1686558fca4ff8ca0b2b70.

That removes the use of the so-called "thread safe" tsafe module, which
adds a lock around many messages, but did not account for the use of
that lock in the paste/httpserver.py module.

An attempt was made to just get rid of that lock, but since there is
limited testing of that area, and few resources to confirm the change,
it's been decided that keeping tsafe and allow the deprecation warning
to be exposed is the best thing to do. With luck the decpration warning
will encourage people to not use the httpserver and choose something
else instead.

Fixes #19