Augment use of imp with importlib.util
imp is deprecated and will go away in Python 3.12 (which is very
soon). The preferred replacement is importlib.

However, importlib is not available in Python 2.7, so we want
to keep both in order to maintain 2.7 some semblance of 2.7
support.

This implementation is based on the importlib docs [1] and passes
the quite robust tests that are already present for the urlparser
module. Unfortunately there's no good way to be 100% certain that
this works for all the many ways that Paste can do an import. So
this patch is mostly hoping that test coverage is sufficient.

[1] https://docs.python.org/3/library/importlib.html#importing-a-source-file-directly

Fixes #75