Cleanup namespace handling in tests
When we did a lot of recent Python2-isms cleanup recently
it broke some tests when run in an environment where a
previous version of paste, or one of the consumers of its
namespace package, were already installed.

This was probably done to quiet warnings. However, it
hides issues with complex environments.

In addition to fixing the problem, this change adds a
tox test environment that should cause a failure if
we introduce the error again.

Fixes #99